+++ assets/js/view.js
@@ -64,6 +64,7 @@ view = {
 					$("#tools_albums, #tools_photo").hide();
 					$("#tools_album").show();
 					album.json.content === false ? $("#button_archive").hide() : $("#button_archive").show();
+					if (lychee.publicMode&&album.json.downloadable==="0") $("#button_archive").hide();
 					if (albumID==="s"||albumID==="f"||albumID==="r") {
 						$("#button_info_album, #button_trash_album, #button_share_album").hide();
 					} else if (albumID==="0") {
@@ -237,12 +238,6 @@ view = {
 
 		},
 
-		description: function() {
-
-			$("#infobox .attr_description").html(album.json.description + " " + build.editIcon("edit_description_album"));
-
-		},
-
 		content: {
 
 			init: function() {
@@ -304,6 +299,12 @@ view = {
 				});
 
 			}
+
+		},
+
+		description: function() {
+
+			$("#infobox .attr_description").html(album.json.description + " " + build.editIcon("edit_description_album"));
 
 		},